requirements.txt
nsepy==0.8
yfinance==0.2.38
kiteconnect==5.0.1
requests==2.31.0
pytz==2024.1
pandas==2.2.1
python-dotenv==1.0.1
